Refactor & Code Cleanup
- Refactor to modularize Mongo & ES Managers as they were previously specific to 'Products'
- Corrected API responses to be more uniform
- Changed ES Manager to raise thrown exceptions
- Added encoder/serializer for cleaner handling of ObjectId Serialization

This should have been broken out into multiple commits in my opinion. That said, the ES & Mongo Managers are now more modular to allow for future additions. Additionally the serialization of ObjectIds from Mongo is much cleaner now and API responses should be in a more consistent and logical pattern.